requirements.txt
behave==1.2.5
decorator==4.0.9
enum34==1.1.2
Flask==0.10.1
Flask-SQLAlchemy==2.1
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
parse==1.6.6
parse-type==0.3.4
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
pycrypto==2.6.1
PyJWT==1.4.0
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.11
traitlets==4.1.0
Werkzeug==0.11.3
wheel==0.24.0
